# server/swagger.yaml
openapi: 3.0.0
info:
  title: RapiDevKit
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /blocks/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Block
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Block
      parameters:
        - name: id
          in: path
          description: Id 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
    post:
      summary: CRUD ACTION update
      tags:
        - Block
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Block"
  /blocks/findByActyve/{key}:
    get:
      summary: CRUD ACTION findByActyve
      tags:
        - Block
      parameters:
        - name: key
          in: path
          description: Id of the resource Actyve to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Block"
  /blocks/findByBlockType/{key}:
    get:
      summary: CRUD ACTION findByBlockType
      tags:
        - Block
      parameters:
        - name: key
          in: path
          description: Id of the resource BlockType to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Block"
  /blocks/findByFontendAPIHook/{key}:
    get:
      summary: CRUD ACTION findByFontendAPIHook
      tags:
        - Block
      parameters:
        - name: key
          in: path
          description: Id of the resource FontendAPIHook to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Block"
  /blocks/{id}/getBlockType:
    get:
      summary: CRUD ACTION getBlockType
      tags:
        - Block
      parameters:
        - name: id
          in: path
          description: ID of Block from BlockType
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
    
  #-----  BlockType  -----
...
  /blocktypes/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - BlockType
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - BlockType
      parameters:
        - name: id
          in: path
          description: Id 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockType"
    post:
      summary: CRUD ACTION update
      tags:
        - BlockType
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockType"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockType"
  /blocktypes/findByName/{key}:
    get:
      summary: CRUD ACTION findByName
      tags:
        - BlockType
      parameters:
        - name: key
          in: path
          description: Id of the resource Name to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockType"
  /blocktypes/{id}/getItemType:
    get:
      summary: CRUD ACTION getItemType
      tags:
        - BlockType
      parameters:
        - name: id
          in: path
          description: ID of BlockType from ItemType
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
    
  #-----  Item  -----
...
  /types/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - ItemType
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - ItemType
      parameters:
        - name: id
          in: path
          description: Id 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemType"
    post:
      summary: CRUD ACTION update
      tags:
        - ItemType
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemType"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemType"
  /types/findByName/{key}:
    get:
      summary: CRUD ACTION findByName
      tags:
        - ItemType
      parameters:
        - name: key
          in: path
          description: Id of the resource Name to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemType"
    
  #-----  Role  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Block:
      properties:
        _id:
          type: string
        Actyve:
          type: boolean
        FontendAPIHook:
          type: string
        Settings:
          type: string
      required:
        -  _id
        - Actyve
    BlockType:
      properties:
        _id:
          type: string
        Code:
          type: string
        Name:
          type: string
      required:
        -  _id
        - Name
...
    ItemType:
      properties:
        _id:
          type: string
        Name:
          type: string
        Schema:
          type: string
        Validation:
          type: string
        Visualization:
          type: string
      required:
        -  _id
...